Take out extra test file in cmd
